cast lookup table threshold back to int, as remote blast expects
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@32884 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/blast/api/blast_options_cxx.cpp
@@ -759,7 +759,7 @@ CBlastOptions::SetWordThreshold(double w)
         m_Local->SetWordThreshold(w);
     }
     if (m_Remote) {
-        m_Remote->SetValue(eBlastOpt_WordThreshold, w);
+        m_Remote->SetValue(eBlastOpt_WordThreshold, static_cast<int>(w));
     }
 }
